+++ BetterCrafting/Models/MatcherIngredient.cs
@@ -19,13 +19,18 @@ public class MatcherIngredient : IIngredient {
 	public readonly Func<Item, bool> ItemMatcher;
 	private readonly (Func<Item, bool>, int)[] IngList;
 
-	public MatcherIngredient(Func<Item, bool> matcher, int quantity, string displayName, Texture2D texture, Rectangle? source = null) {
+	private readonly Func<string> _displayName;
+	private readonly Func<Texture2D> _texture;
+
+	private Rectangle? _source;
+
+	public MatcherIngredient(Func<Item, bool> matcher, int quantity, Func<string> displayName, Func<Texture2D> texture, Rectangle? source = null) {
 		ItemMatcher = matcher;
 		Quantity = quantity;
 
-		DisplayName = displayName;
-		Texture = texture;
-		SourceRectangle = source ?? texture.Bounds;
+		_displayName = displayName;
+		_texture = texture;
+		_source = source;
 
 		IngList = new (Func<Item, bool>, int)[] {
 			(ItemMatcher, Quantity)
@@ -36,11 +41,17 @@ public class MatcherIngredient : IIngredient {
 
 	public bool SupportsQuality => true;
 
-	public string DisplayName { get; }
+	public string DisplayName => _displayName();
 
-	public Texture2D Texture { get; }
+	public Texture2D Texture => _texture();
 
-	public Rectangle SourceRectangle { get; }
+	public Rectangle SourceRectangle {
+		get {
+			if (!_source.HasValue)
+				_source = _texture().Bounds;
+			return _source.Value;
+		}
+	}
 
 	public int Quantity { get; }